pyproject.toml
[project]
name = "course-consultant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.8.0",
    "flask>=3.0.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
]
